[cli-cmd] Fix return type
time_t is signed on UNIX/POSIX compliant systems.

+++ cli/src/cli.h
@@ -141,7 +141,7 @@ struct cli_state {
     char *remote_host;
     int remote_port;
     int mode;
-    int await_connected;
+    int64_t await_connected;
 
     time_t default_conn_timeout;
